Create teamInput and adjust gameInput to be scoreInput
