+++ src/__tests__/HomeElems.test.js
@@ -8,9 +8,9 @@ describe('The Home page has ', () => {
   beforeEach(() => {
     renderedComponent = render(<Home />);
   });
-  test('a div with class "home"', () => {
+  test('a main elem with class "home"', () => {
     const { container } = renderedComponent;
-    const appElement = container.querySelector('div');
+    const appElement = container.querySelector('main');
     expect(appElement).toHaveClass('home');
   });
   test('an element with class "score-sheet"', () => {
